project/requirements.txt
asgiref==3.8.1 
certifi==2025.6.15 
charset-normalizer==3.4.2 
Django==5.1.1 
django-cors-headers==4.4.0 
django-filter==24.3 
djangorestframework==3.15.2 
djangorestframework-simplejwt==5.3.1 
Faker==29.0.0 
idna==3.10 
pillow==10.4.0 
razorpay==1.4.2 
rest-framework-simplejwt==0.0.2 
six==1.16.0 
sqlparse==0.5.1 
tzdata==2024.1 
urllib3==2.4.0
